HSEARCH-3517 Use a long for totalHitCountThreshold

+++ backend/elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/search/query/impl/ElasticsearchSearchQueryBuilder.java
@@ -72,7 +72,7 @@ public class ElasticsearchSearchQueryBuilder<H>
 	private Long timeoutValue;
 	private TimeUnit timeoutUnit;
 	private boolean exceptionOnTimeout;
-	private Integer totalHitCountThreshold;
+	private Long totalHitCountThreshold;
 	private ElasticsearchSearchRequestTransformer requestTransformer;
 
 	public ElasticsearchSearchQueryBuilder(
@@ -125,7 +125,7 @@ public class ElasticsearchSearchQueryBuilder<H>
 	}
 
 	@Override
-	public void totalHitCountThreshold(int totalHitCountThreshold) {
+	public void totalHitCountThreshold(long totalHitCountThreshold) {
 		this.totalHitCountThreshold = totalHitCountThreshold;
 	}
